requirements.txt
Flask
Flask-Cors
Flask-SQLAlchemy
tensorflow
opencv-python
numpy
Pillow
